demonstration of animation with a sprite reel
Exploding a GIF out into a reel is a tedious process. If I need to do
that for a game, I'll want to make a tool to do that for me. It would
also be nice to devise a method of storing the frame size along with the
sprite reel itself.